Revert "chained function calls via consolidating functionality from Command and YaspClass into a single class"
unfortunately this commit introduces a class boundary at every command that functions can't cross. originally, due to YaspClass managing the tree (instead of the individual nodes managing themselves) and calling instance_eval on _itself_ with a new head, we would avoid nesting classes for the most part - user defined classes were an exception, but you shouldnt expect your globally scoped function to work inside of a class. nesting classes introduces a boundary that functions cannot cross, and method chaining is not as important as that imo

This reverts commit d7b4829b932ea7814587b82cab859be4940dfb4c.